requirements.txt
Django==2.1.7
djangorestframework==3.9.2
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
Pygments==2.3.1
pytz==2018.9
